fix(#2764 R9): pinned/external isolation, durable child ledger, TUI/MCP rollback, opaque tokens
Addresses the R8 rejection (root verdict + independent supplemental audits):

1. Pinned managed delete STRUCTURALLY excludes external/name-only removal:
   handle_delete's external arm runs only for unpinned deletes — a same-name
   external coexisting with pinned managed A is never deregistered in A's
   stead (that returned stopped:true without stopping A).

2. Missing exact registry handle is UNKNOWN, not terminal proof: new durable
   child-lifecycle ledger (agent::child_ledger, runtime/child-pids/<uuid>.json)
   written at spawn registration (managed agents), cleared only on PROVEN
   terminal exit. delete_transaction_expecting with no handle consults it +
   an OS liveness probe: no entry = durable negative (never spawned/already
   proven) → terminal; recorded pid provably gone → terminal + ledger
   cleared; recorded pid alive/unverifiable → UNKNOWN → stopped:false,
   nothing released, never killing an unverified pid (pid reuse).

3. Fugu owns its CreateAdmission (refusal aborts — no more warn-and-continue
   into pane creation), mirroring the Backend arm.

4. Hold-through-success-or-rollback on every guarded TUI/MCP entry:
   tui_spawn::rollback_created_instance undoes fleet entry + topic mapping on
   pane-create failure (menu Backend/Fugu + palette, while the admission is
   still held); the MCP create rollback now also removes the worktree and the
   working dir THIS call materialized (pre-existing operator dirs are never
   touched). Boot/crash post-spawn rollback already routes through
   SpawnRollback/delete_transaction; their admission holds landed in R8.

5. Re-entry tokens are opaque and unguessable: random 128-bit (uuid v4)
   values carried as strings through the SPAWN RPC — the sequential u64 was
   forgeable by a same-UID caller driving SPAWN directly.

REDs: pinned_delete_never_removes_same_name_external; lost-child alive →
UNKNOWN (ledger + port retained) and lost-child provably-dead → terminal +
ledger cleared; A→B pinned no-retarget re-justified via durable-negative
proof; TUI rollback removes fleet+topic; MCP rollback removes created
workdir + keeps pre-existing operator dir.


Agend-Agent: claude-aef7c0
Agend-Task: t-20260713071508674730-70517-17
Agend-Branch: fix/2764-owned-workspace-cleanup
Agend-Issued-At: 2026-07-13T07:16:29.137602+00:00

+++ src/agent_ops.rs
@@ -540,7 +540,7 @@ pub fn spawn_one(
     work_dir: &Path,
     size: (u16, u16),
     env: Option<&std::collections::HashMap<String, String>>,
-    create_token: Option<u64>,
+    create_token: Option<u128>,
 ) -> anyhow::Result<crate::backend::SpawnMode> {
     // #2764 R7/R8 (codex P0-1): admission BEFORE the first side effect — the
     // dir creation + skills install below run long before spawn_agent's
